feat: Create second funcionality
With this feature added the user can now choose the number of squares
per row that will be created on the sketch area.
If the user decides not to input anything, the rowSizeSelection event
listener is never triggered. To avoid that was added a default value to
the rowSize variable of 16.